requirements.txt
streamlit
pandas
faiss-cpu
langchain
transformers
sentence-transformers
Pillow
